@@ -1,6 +1,6 @@
 {
   "name": "svelte-redux-adapter",
-  "version": "1.0.3",
+  "version": "1.0.4",
   "author": {
     "name": "Cola Fanta Wlk <colafantawlk@gmail.com>"
   },
@@ -43,33 +43,33 @@
     "dist"
   ],
   "peerDependencies": {
-    "svelte": "^3.54.0"
+    "svelte": "^4.2.0"
   },
   "devDependencies": {
-    "@fontsource/fira-mono": "^4.5.10",
-    "@playwright/test": "^1.33.0",
+    "@fontsource/fira-mono": "^5.0.8",
+    "@playwright/test": "^1.37.1",
     "@reduxjs/toolkit": "^1.9.5",
-    "@sveltejs/adapter-auto": "^2.0.1",
-    "@sveltejs/kit": "^1.16.0",
-    "@sveltejs/package": "^2.0.2",
-    "@trivago/prettier-plugin-sort-imports": "^4.1.1",
+    "@sveltejs/adapter-auto": "^2.1.0",
+    "@sveltejs/kit": "^1.23.0",
+    "@sveltejs/package": "^2.2.2",
+    "@trivago/prettier-plugin-sort-imports": "^4.2.0",
     "@types/cookie": "^0.5.1",
-    "@typescript-eslint/eslint-plugin": "^5.59.2",
-    "@typescript-eslint/parser": "^5.59.2",
-    "@vitest/coverage-c8": "^0.31.0",
-    "eslint": "^8.39.0",
-    "eslint-config-prettier": "^8.8.0",
+    "@typescript-eslint/eslint-plugin": "^6.4.1",
+    "@typescript-eslint/parser": "^6.4.1",
+    "@vitest/coverage-c8": "^0.33.0",
+    "eslint": "^8.48.0",
+    "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-svelte3": "^4.0.0",
-    "prettier": "^2.8.8",
-    "prettier-plugin-svelte": "^2.10.0",
-    "publint": "^0.1.11",
-    "svelte": "^3.58.0",
-    "svelte-check": "^3.3.0",
-    "svelte-preprocess": "^5.0.3",
-    "tslib": "^2.5.0",
-    "typescript": "^5.0.4",
-    "vite": "^4.3.4",
-    "vitest": "^0.31.0"
+    "prettier": "^3.0.2",
+    "prettier-plugin-svelte": "^3.0.3",
+    "publint": "^0.2.2",
+    "svelte": "^4.2.0",
+    "svelte-check": "^3.5.0",
+    "svelte-preprocess": "^5.0.4",
+    "tslib": "^2.6.2",
+    "typescript": "^5.2.2",
+    "vite": "^4.4.9",
+    "vitest": "^0.34.3"
   },
   "svelte": "./dist/index.js",
   "types": "./dist/index.d.ts",
